cleanup up debug file

+++ pypy/lang/gameboy/timer.py
@@ -156,5 +156,5 @@ class Clock(object):
         pass
     
     def get_time(self):
-        return int(time.time()*1000)
+        return (time.time()*1000)
